+++ activestorage/app/controllers/active_storage/direct_uploads_controller.rb
@@ -15,9 +15,6 @@ class ActiveStorage::DirectUploadsController < ActiveStorage::BaseController
     end
 
     def direct_upload_json(blob)
-      blob.as_json(root: false, methods: :signed_id).merge(direct_upload: {
-        url: blob.service_url_for_direct_upload,
-        headers: blob.service_headers_for_direct_upload
-      })
+      blob.as_json(root: false, methods: [ :signed_id, :direct_upload ])
     end
 end