+++ activestorage/lib/active_storage.rb
@@ -39,6 +39,7 @@ module ActiveStorage
   autoload :Service
   autoload :Previewer
   autoload :Analyzer
+  autoload :DirectUpload
 
   mattr_accessor :logger
   mattr_accessor :verifier