+++ activestorage/lib/active_storage/service.rb
@@ -104,12 +104,21 @@ module ActiveStorage
       raise NotImplementedError
     end
 
+    def direct_upload(blob)
+      ActiveStorage::DirectUpload.new(blob)
+    end
+
     # Returns a signed, temporary URL that a direct upload file can be PUT to on the +key+.
     # The URL will be valid for the amount of seconds specified in +expires_in+.
     # You must also provide the +content_type+, +content_length+, and +checksum+ of the file
     # that will be uploaded. All these attributes will be validated by the service upon upload.
     def url_for_direct_upload(key, expires_in:, content_type:, content_length:, checksum:)
       raise NotImplementedError
+    end
+
+    # Returns http method for +url_for_direct_upload+ requests.
+    def method_for_direct_upload
+      'PUT'
     end
 
     # Returns a Hash of headers for +url_for_direct_upload+ requests.